refactor: 라우터 lint 오류 해결
