+++ client/.eslintrc.json
@@ -13,6 +13,7 @@
     "project": ["client/tsconfig.json"]
   },
   "rules": {
-    "react/prop-types": "off"
+    "react/prop-types": "off",
+    "@typescript-eslint/no-empty-interface": "off"
   }
 }
