+++ client/public/index.html
@@ -5,6 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>배민문구사</title>
+    <base href="/" />
   </head>
   <body>
     <div id="root"></div>
