+++ client/src/lib/router/components/link.tsx
@@ -1,12 +1,12 @@
 import React, { FC, MouseEvent } from 'react';
-import { useHistory } from '../hooks/use-history';
+import useHistory from '../hooks/use-history';
 
 interface ILink {
   to: string;
   replace?: boolean;
 }
 
-export const Link: FC<ILink> = ({ children, to, replace: isReplace }) => {
+const Link: FC<ILink> = ({ children, to, replace: isReplace }) => {
   const { push, replace } = useHistory();
 
   const onClick = (e: MouseEvent) => {
@@ -24,3 +24,5 @@ export const Link: FC<ILink> = ({ children, to, replace: isReplace }) => {
     </a>
   );
 };
+
+export default Link;
