+++ client/src/lib/router/context/route-context.ts
@@ -1,5 +1,5 @@
 import { createContext } from 'react';
 
-interface IRouteContext extends Record<string, unknown> {}
+export interface IRouteContext {}
 
 export const RouteContext = createContext<IRouteContext>({});
