+++ client/src/lib/router/index.ts
@@ -1,10 +1,7 @@
-export { BrowserRouter } from './components/browser-router';
-export { Route } from './components/route';
-export { Switch } from './components/switch';
-export { Link } from './components/link';
+export { default as BrowserRouter } from './components/browser-router';
+export { default as Route } from './components/route';
+export { default as Switch } from './components/switch';
+export { default as Link } from './components/link';
 
-export { useParams } from './hooks/use-params';
-export { useHistory } from './hooks/use-history';
-
-export { RouterContext } from './context/router-context';
-export { RouteContext } from './context/route-context';
+export { default as useParams } from './hooks/use-params';
+export { default as useHistory } from './hooks/use-history';
